startup_gateway/requirements-dev.txt
pytest==8.4.0
pytest-django==4.5.2
pytest-cov==4.1.0


